citi/requirements.txt
scrapy==2.11.0
pandas==2.2.1
boto3==1.28.40
fastparquet==2024.5.0
openpyxl==3.1.2
pdfplumber==0.11.0
tabula-py==2.9.0
html5lib==1.1
bs4==0.0.2
slack_sdk==3.27.2